+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog for cfgrib
 0.9.6 (unreleased)
 ------------------
 
-- Add read-only support on *Windows* after you install *ecCodes* on *conda*.
+- Add support for *Windows* by installing *ecCodes* via *conda*.
   See: `#7 <https://github.com/ecmwf/cfgrib/issues/7>`_
 - Do not set explicit timezone in ``units`` to avoid crashing some versions of xarray.
   See: `#44 <https://github.com/ecmwf/cfgrib/issues/44>`_.